Movido para o use case
