+++ src/SME.SGP.Aplicacao/CasosDeUso/AtividadeAvaliativa/ObterAtividadesNotasAlunoPorTurmaPeriodoUseCase.cs
@@ -2,6 +2,7 @@
 using SME.SGP.Aplicacao.Interfaces;
 using SME.SGP.Infra;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace SME.SGP.Aplicacao
@@ -12,7 +13,32 @@ namespace SME.SGP.Aplicacao
         {
         }
 
-        public Task<IEnumerable<AvaliacaoNotaAlunoDto>> Executar(FiltroTurmaAlunoPeriodoEscolarDto param)
-            => mediator.Send(new ObterAtividadesNotasAlunoPorTurmaPeriodoQuery(param.TurmaId, param.PeriodoEscolarId, param.AlunoCodigo, param.ComponenteCurricular));
+        public async Task<IEnumerable<AvaliacaoNotaAlunoDto>> Executar(FiltroTurmaAlunoPeriodoEscolarDto param)
+        {
+          var retorno =  (await mediator.Send(new ObterAtividadesNotasAlunoPorTurmaPeriodoQuery(param.TurmaId, param.PeriodoEscolarId, param.AlunoCodigo, param.ComponenteCurricular))).ToList();
+
+          retorno = await ObterAusencia(param, retorno);
+          return retorno;
+        }
+        
+        private async Task<List<AvaliacaoNotaAlunoDto>> ObterAusencia(FiltroTurmaAlunoPeriodoEscolarDto request, List<AvaliacaoNotaAlunoDto> listAtividades)
+        {
+            var retorno = new List<AvaliacaoNotaAlunoDto>();
+            var datasDasAtividadesAvaliativas = retorno.Select(x => x.Data).ToArray();
+            var turma = await mediator.Send(new ObterTurmaPorIdQuery(request.TurmaId));
+
+            var ausenciasDasAtividadesAvaliativas = (await mediator.Send(new ObterAusenciasDaAtividadesAvaliativasPorAlunoQuery(turma.CodigoTurma, datasDasAtividadesAvaliativas, request.ComponenteCurricular, request.AlunoCodigo))).ToList();
+
+            foreach (var atividade in listAtividades)
+            {
+                var ausente = ausenciasDasAtividadesAvaliativas
+                    .Any(a => a.AlunoCodigo == request.AlunoCodigo && a.AulaData.Date == atividade.Data.Date);
+
+                atividade.Ausente = ausente;
+                retorno.Add(atividade);
+            }
+
+            return retorno;
+        }
     }
 }
